+++ spec/support/shared_examples/common_calculations_shared_examples.rb
